+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # v2.14.17
 
 * add `default_return`
+* added `timeout` and `timeouted` property to `satella.time.measure`
+* some methods from `measure` are pending deprecation